SimpleTextField 새 파일로 분리
