+++ presentation/src/main/java/com/pocs/presentation/view/component/textfield/SimpleTextField.kt
@@ -1,0 +1,39 @@
+package com.pocs.presentation.view.component.textfield
+
+import androidx.compose.foundation.text.KeyboardOptions
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextField
+import androidx.compose.material3.TextFieldDefaults
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun SimpleTextField(
+    hint: String,
+    value: String,
+    maxLength: Int,
+    keyboardOptions: KeyboardOptions = KeyboardOptions.Default,
+    onValueChange: (String) -> Unit,
+    modifier: Modifier
+) {
+    TextField(
+        colors = TextFieldDefaults.textFieldColors(
+            focusedIndicatorColor = Color.Transparent,
+            disabledIndicatorColor = Color.Transparent,
+            unfocusedIndicatorColor = Color.Transparent,
+            containerColor = Color.Transparent
+        ),
+        keyboardOptions = keyboardOptions,
+        placeholder = { Text(hint) },
+        value = value,
+        onValueChange = {
+            if (it.length <= maxLength) {
+                onValueChange(it)
+            }
+        },
+        modifier = modifier
+    )
+}
